Computer-Use/requirements.txt
anthropic
pyautogui
Pillow
python-dotenv
pyperclip
